requirements.txt
Flask==2.3.3
pandas==2.1.4
numpy==1.26.4
scikit-learn==1.4.2
geojson==3.0.1
gunicorn==21.2.0
boto3==1.34.34
joblib==1.3.2
xgboost==2.0.3
PyYAML==6.0.1